Move flux point core tests over
